+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Allow to run classifier and regressor with multiple models [#557](https://github.com/PriorLabs/TabPFN/pull/557).
 - Disable all telemetry in unit tests.
 
 ## [2.2.1] - 2025-09-17